Snapshot R2DBC connection configuration
